source/api/neptune-integrator/api/item/get/get-item-by-location-api.ts
// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
// SPDX-License-Identifier: Apache-2.0

import { ApiInterface, apiResponses } from "api/api-interface";
import { ApiResponse } from "api/api-response";
import { NeptuneDB } from "neptune/db/neptune-db";
import { EdgeDirection, VertexLabel } from "neptune/model/constants";
import Logger from "util/logger";
import { PathParser } from "util/pathParser";

export class GetItemByLocationApi implements ApiInterface {
    logger: Logger;
    static EXPECTED_PATH_PARTS = 4;
    db: NeptuneDB;

    constructor(db: NeptuneDB) {
        this.logger = new Logger(this.constructor.name);
        this.db = db;
    }

    async execute(path: string, _body: string | null, _headers: object): Promise<ApiResponse> {
        this.logger.info(`Path: ` + path);
        const locationId = PathParser.getLastItemInPath(path, GetItemByLocationApi.EXPECTED_PATH_PARTS);
        const items = await this.db.getAllConnected(
            VertexLabel.LOCATION,
            VertexLabel.ITEM,
            locationId,
            EdgeDirection.IN
        );
        if (items !== undefined) {
            return apiResponses._200({ items: items });
        }
        return apiResponses._404();
    }
}
